+++ app/src/test/java/com/growit/app/retrospect/controller/RetrospectControllerTest.java
@@ -17,7 +17,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.growit.app.common.TestSecurityUtil;
 import com.growit.app.fake.goal.PlanFixture;
 import com.growit.app.fake.retrospect.RetrospectFixture;
-import com.growit.app.fake.todo.ToDoFixture;
 import com.growit.app.goal.domain.goal.plan.Plan;
 import com.growit.app.retrospect.controller.dto.request.CreateRetrospectRequest;
 import com.growit.app.retrospect.controller.dto.request.UpdateRetrospectRequest;
@@ -26,10 +25,6 @@ import com.growit.app.retrospect.domain.retrospect.dto.RetrospectWithPlan;
 import com.growit.app.retrospect.usecase.CreateRetrospectUseCase;
 import com.growit.app.retrospect.usecase.GetRetrospectUseCase;
 import com.growit.app.retrospect.usecase.UpdateRetrospectUseCase;
-import com.growit.app.todo.domain.ToDo;
-import com.growit.app.todo.usecase.GetTodayMissionUseCase;
-import java.time.LocalDate;
-import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -57,7 +52,6 @@ class RetrospectControllerTest {
   @MockitoBean private CreateRetrospectUseCase createRetrospectUseCase;
   @MockitoBean private UpdateRetrospectUseCase updateRetrospectUseCase;
   @MockitoBean private GetRetrospectUseCase getRetrospectUseCase;
-  @MockitoBean private GetTodayMissionUseCase getTodayMissionUseCase;
 
   @BeforeEach
   void setUp(
@@ -166,44 +160,6 @@ class RetrospectControllerTest {
                             fieldWithPath("data.plan.weekOfMonth").description("계획 주차"),
                             fieldWithPath("data.plan.content").description("계획 내용"),
                             fieldWithPath("data.content").description("회고 내용"))
-                        .build())));
-  }
-
-  @Test
-  void getTodayMission() throws Exception {
-    // given
-    List<ToDo> todoList =
-        List.of(
-            ToDoFixture.customToDo("id", "user-1", LocalDate.now(), "planId", "goalId"),
-            ToDoFixture.customToDo("id2", "user-1", LocalDate.now(), "planId", "goalId"));
-    given(getTodayMissionUseCase.execute(any(), any())).willReturn(todoList);
-
-    // when & then
-    mockMvc
-        .perform(
-            get("/todos/today-mission/date")
-                .header("Authorization", "Bearer mock-jwt-token")
-                .contentType(MediaType.APPLICATION_JSON))
-        .andExpect(status().isOk())
-        .andDo(
-            document(
-                "get-today-mission",
-                preprocessRequest(prettyPrint()),
-                preprocessResponse(prettyPrint()),
-                resource(
-                    new ResourceSnippetParametersBuilder()
-                        .tag("Todos")
-                        .summary("오늘 미션 조회")
-                        .description("오늘 날짜의 미완료 ToDo 리스트를 조회합니다.")
-                        .responseFields(
-                            fieldWithPath("data[].id").type(STRING).description("TODO ID"),
-                            fieldWithPath("data[].goalId").type(STRING).description("목표 ID"),
-                            fieldWithPath("data[].planId").type(STRING).description("계획 ID"),
-                            fieldWithPath("data[].date").type(STRING).description("할 일 날짜"),
-                            fieldWithPath("data[].content").type(STRING).description("내용"),
-                            fieldWithPath("data[].isCompleted")
-                                .type("Boolean")
-                                .description("완료 여부"))
                         .build())));
   }
 }
